Move RPCSendTransaction tests to node/rpc_test.go
